backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
pandas==2.1.3
openpyxl==3.1.2
supabase==2.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
# Remove specific httpx version to let pip resolve compatibility
httpx>=0.24.0,<0.25.0
